+++ @pauliescanlon/gatsby-theme-terminal/gatsby-ssr.js
@@ -1,5 +1,5 @@
 const React = require('react')
-const { PageElement } = require('./src/components/PageElement')
+const { PageElement } = require('./src/components/page-element')
 
 exports.wrapPageElement = ({ element }) => {
   return <PageElement>{element}</PageElement>
