+++ @pauliescanlon/gatsby-theme-terminal/src/components/Logo/index.js
@@ -1,1 +1,1 @@
-export { Logo } from './Logo'
+export { Logo } from './logo'
