+++ @pauliescanlon/gatsby-theme-terminal/src/components/Main/index.js
@@ -1,1 +1,1 @@
-export { Main } from './Main'
+export { Main } from './main'
