+++ @pauliescanlon/gatsby-theme-terminal/src/components/Nav/index.js
@@ -1,1 +1,1 @@
-export { Nav } from './Nav'
+export { Nav } from './nav'
