--- @pauliescanlon/gatsby-theme-terminal/src/components/PageElement/index.js
@@ -1,1 +1,0 @@
-export { PageElement } from './PageElement'
