--- @pauliescanlon/gatsby-theme-terminal/src/components/SiteMetaData/index.js
@@ -1,1 +1,0 @@
-export { SiteMetaData } from './SiteMetaData'
