--- @pauliescanlon/gatsby-theme-terminal/src/components/SourceArticle/index.js
@@ -1,1 +1,0 @@
-export { SourceArticle } from './SourceArticle'
