--- @pauliescanlon/gatsby-theme-terminal/src/components/SourceDays/index.js
@@ -1,1 +1,0 @@
-export { SourceDays } from './SourceDays'
