--- @pauliescanlon/gatsby-theme-terminal/src/components/SourceList/index.js
@@ -1,1 +1,0 @@
-export { SourceList } from './SourceList'
