--- @pauliescanlon/gatsby-theme-terminal/src/components/SourceMonths/index.js
@@ -1,1 +1,0 @@
-export { SourceMonths } from './SourceMonths'
