--- @pauliescanlon/gatsby-theme-terminal/src/components/SourceTags/index.js
@@ -1,1 +1,0 @@
-export { SourceTags } from './SourceTags'
