--- @pauliescanlon/gatsby-theme-terminal/src/components/SourceWords/index.js
@@ -1,1 +1,0 @@
-export { SourceWords } from './SourceWords'
