+++ @pauliescanlon/gatsby-theme-terminal/src/components/page-element/index.js
@@ -1,0 +1,1 @@
+export { PageElement } from './page-element'
