+++ @pauliescanlon/gatsby-theme-terminal/src/components/page-element/page-element.js
