+++ @pauliescanlon/gatsby-theme-terminal/src/components/site-metadata/index.js
@@ -1,0 +1,1 @@
+export { SiteMetaData } from './site-metadata'
