+++ @pauliescanlon/gatsby-theme-terminal/src/components/site-metadata/site-metadata.js
