+++ @pauliescanlon/gatsby-theme-terminal/src/components/source-article/index.js
@@ -1,0 +1,1 @@
+export { SourceArticle } from './source-article'
