+++ @pauliescanlon/gatsby-theme-terminal/src/components/source-article/source-article.js
