+++ @pauliescanlon/gatsby-theme-terminal/src/components/source-days/index.js
@@ -1,0 +1,1 @@
+export { SourceDays } from './source-days'
