+++ @pauliescanlon/gatsby-theme-terminal/src/components/source-list/index.js
@@ -1,0 +1,1 @@
+export { SourceList } from './source-list'
