+++ @pauliescanlon/gatsby-theme-terminal/src/components/source-months/index.js
@@ -1,0 +1,1 @@
+export { SourceMonths } from './source-months'
