+++ @pauliescanlon/gatsby-theme-terminal/src/components/source-tags/index.js
@@ -1,0 +1,1 @@
+export { SourceTags } from './source-tags'
