+++ @pauliescanlon/gatsby-theme-terminal/src/components/source-words/index.js
@@ -1,0 +1,1 @@
+export { SourceWords } from './source-words'
