+++ @pauliescanlon/gatsby-theme-terminal/src/components/source-words/source-words.js
