+++ @pauliescanlon/gatsby-theme-terminal/src/data/index.js
@@ -1,6 +1,6 @@
-export { useConfig } from './useConfig'
-export { useNavigation } from './useNavigation'
-export { useDates } from './useDates'
-export { useTags } from './useTags'
-export { useWords } from './useWords'
-export { useSource } from './useSource'
+export { useConfig } from './use-config'
+export { useNavigation } from './use-navigation'
+export { useDates } from './use-dates'
+export { useTags } from './use-tags'
+export { useWords } from './use-words'
+export { useSource } from './use-source'
